Fix WOL lobby crash: 64-bit pointer truncation in GadgetListBox
Root cause: GLM_GET_SELECTION handler in GadgetListBox.cpp was casting
list->selections (Int* pointer) to Int (32-bit), truncating the upper
32 bits of the 64-bit address on arm64/macOS. The truncated pointer was
then dereferenced in PopulateLobbyPlayerListbox(), causing SIGSEGV.

Fix: Use intptr_t cast to preserve the full 64-bit pointer width when
passing the selections array through the window message system.

Also includes:
- Thread-safety: copy player/buddy maps before iteration in
  PopulateLobbyPlayerListbox to prevent concurrent modification crashes
- CRC: use GeneralsOnlineZH_60.exe CRC (0xDA2A0068) as base for fake
  EXE CRC to match GenTool Online clients

+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/Gadget/GadgetListBox.cpp
@@ -1792,7 +1792,7 @@ WindowMsgHandledType GadgetListBoxSystem( GameWindow *window, UnsignedInt msg,
 		{
 
 			if( list->multiSelect )
-				*(Int*)mData2 = (Int)(intptr_t)list->selections;
+				*(intptr_t*)mData2 = (intptr_t)list->selections;
 			else
 				*(Int*)mData2 = list->selectPos;
